Create Chronos as Maven Multi-Module Project

+++ install/install.sh
@@ -1,12 +1,15 @@
 # Versions
 SIGNAL_VERSION=0.13.4
 JAVA_VERSION=21.0.2-tem
-MYSQL_VERSION=0.8.29-1
+MYSQL_VERSION=0.8.32-1
 
 # Create Directories
 mkdir /opt/chronos
 mkdir /opt/chronos/logs
 mkdir /opt/chronos/logs/cron
+
+apt install curl
+apt install zip
 
 # Install Java
 curl -s "https://get.sdkman.io" | bash
@@ -15,11 +18,19 @@ sdk install java $JAVA_VERSION
 # Install MySQL
 curl -OL https://dev.mysql.com/get/mysql-apt-config_"$MYSQL_VERSION"_all.deb
 apt install ./mysql-apt-config_"$MYSQL_VERSION"_all.deb
+apt update
 apt install mysql-server
+rm mysql-apt-config_"$MYSQL_VERSION"_all.deb
 # Install Signal CLI
-wget https://github.com/AsamK/signal-cli/releases/download/v"${SIGNAL_VERSION}"/signal-cli-"${SIGNAL_VERSION}".tar.gz
-tar xf signal-cli-"${SIGNAL_VERSION}".tar.gz -C /opt
-ln -sf /opt/signal-cli-"${SIGNAL_VERSION}"/bin/signal-cli /usr/local/bin/
+if [ -e /usr/local/bin/signal-cli ]
+then
+  echo "Signal CLI is already installed, skipping"
+else
+  wget https://github.com/AsamK/signal-cli/releases/download/v"${SIGNAL_VERSION}"/signal-cli-"${SIGNAL_VERSION}".tar.gz
+  tar xf signal-cli-"${SIGNAL_VERSION}".tar.gz -C /opt
+  ln -sf /opt/signal-cli-"${SIGNAL_VERSION}"/bin/signal-cli /usr/local/bin/
+  rm signal-cli-"${SIGNAL_VERSION}".tar.gz
+fi
 # Install jq
 apt-get install jq -y
 
@@ -32,7 +43,7 @@ mysql -u root --execute="CREATE USER 'chronos'@'localhost' IDENTIFIED BY '""$PAS
 mysql -u root --execute="GRANT ALL ON *.* TO 'chronos'@'localhost';"
 
 # Set Chronos URL
-echo "Please enter the URL Chronos shall be reachable with:"
+echo "Please enter the URL Chronos shall be reachable with (including http(s) and port if necessary:"
 read -r url
 echo chronos.base-url="$url" | tee /opt/chronos/chronos.properties
 
@@ -46,11 +57,6 @@ line="02 * * * * /root/.sdkman/candidates/java/$JAVA_VERSION/bin/java -jar /opt/
 
 # Install chronos
 cp ./update.sh /opt/chronos/
-cp ./update-cron.sh /opt/chronos/
-cp ./update-frontend.sh /opt/chronos/
 cd /opt/chronos || exit
 chmod +x ./update.sh
-chmod +x ./update-cron.sh
-chmod +x ./update-frontend.sh
 ./update.sh
-systemctl start chronos
